pyproject.toml
[project]
name = "multi-agents-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "aiosqlite>=0.21.0",
    "azure-identity>=1.21.0",
    "azure-keyvault>=4.2.0",
    "fastapi>=0.115.11",
    "geopy>=2.4.1",
    "greenlet>=3.1.1",
    "motor>=3.7.0",
    "openai>=1.66.3",
    "openai-agents>=0.0.9",
    "openai-messages-token-helper>=0.1.11",
    "greenlet>=3.1.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
    "pytz>=2025.2",
    "azure-cognitiveservices-speech>=1.43.0",
    "arize-phoenix>=8.29.0",
    "openinference-instrumentation-openai>=0.1.27",
    "openpyxl>=3.1.5",
]
